feat: add analytics event handling for qr code
