+++ core/plugins/rybbit.ts
@@ -3,14 +3,15 @@ import rybbit from '@rybbit/js'
 export default defineNuxtPlugin(() => {
   if (import.meta.client) {
     rybbit.init({
-      analyticsHost: 'https://analytics.mindenit.org/api/script.js',
+      analyticsHost: 'https://analytics.mindenit.org/api',
       siteId: '2',
+      debug: process.env.NODE_ENV !== 'production',
     })
   }
 
   return {
     provide: {
-      rybbit: rybbit,
+      rybbit,
     },
   }
 })
